+++ commands/users/elo.js
@@ -3,6 +3,7 @@
 // self-defined helper functions
 const { ApplicationCommandOptionType } = require('discord.js');
 const { getScore, updateScoreOnce } = require('../../helper_functions/db_helper.js');
+const { reply } = require('../../helper_functions/comm_helper.js');
 const helper = require('../../helper_functions/helper.js');
 
 module.exports = {
@@ -55,7 +56,7 @@ module.exports = {
         const user_id = user.id;
 
         const user_option = interaction.options.getString('option');
-        console.log(`/elo command is ${user_option} by ${interaction.user.username} for ${user.username}`);
+        console.log(`elo: requesting ${user_option} by ${interaction.user.username} for ${user.username}`);
 
         // if get, return elo of the user from database
         if (user_option === 'get') {
@@ -63,8 +64,8 @@ module.exports = {
             // if rank doesnt exists, throw error and react accordingly
             let score;
             if ((score = getScore(user_id)) === undefined) {
-                console.log(`invalid user elo for user ${user.username}`);
-                interaction.reply(`Error: ${user} has no recorded elo. Set ${user}'s elo before before trying to get it`);
+                console.log(`elo: invalid user elo for user ${user.username}`);
+                reply(interaction, `Error: ${user} has no recorded elo. Set ${user}'s elo before before trying to get it`);
                 return undefined;
             }
 
@@ -72,10 +73,10 @@ module.exports = {
             let elo;
             if ((elo = helper.scoreToElo(score)) === undefined) {
                 // this error should never happen - previous error checking should resolve this
-                interaction.reply(`Error: this is an unexpected error 🤔 please contact ${client.my_maker} with ERRONO 1, and set ${user}'s elo again`);
+                reply(interaction, `Error: this is an unexpected error 🤔 please contact ${client.my_maker} with ERRONO 1, and set ${user}'s elo again`);
                 return undefined;
             }
-            interaction.reply(elo);
+            reply(interaction, elo);
         }
         // if set, change value of user (unless attempting to change someone else's rank and not admin)
         else if (user_option === 'set') {
@@ -87,20 +88,20 @@ module.exports = {
             // ensure that elo is provided
             let elo;
             if ((elo = interaction.options.getString('elo')) === null) {
-                interaction.reply('Error: Elo is required for set function');
+                reply(interaction, 'Error: Elo is required for set function');
                 return undefined;
             }
 
             // calculate the score based on the elo provided
             let score;
             if ((score = helper.eloToScore(elo)) === -1) { // if -1, then error, so return
-                interaction.reply('Error: invalid elo obtained, so cannot process. Please try again');
+                reply(interaction, 'Error: invalid elo obtained, so cannot process. Please try again');
                 return undefined;
             }
 
             // add data to temp database
             updateScoreOnce(user_id, score);
-            interaction.reply(`updating ${user}'s rank to ${elo}`);
+            reply(interaction, `updating ${user}'s rank to ${elo}`);
         }
     },
 };