+++ commands/users/reroll.js
@@ -2,7 +2,7 @@
 
 // self-defined helper functions
 const { readData } = require('../../helper_functions/db_helper.js');
-const { reply } = require('../../helper_functions/event_helper.js');
+const { reply } = require('../../helper_functions/comm_helper.js');
 const helper = require('../../helper_functions/helper.js');
 
 module.exports = {
@@ -21,7 +21,7 @@ module.exports = {
         const data = readData();
 
         if (Object.entries(data.cached_players).length === 0) { // check if cached team is empty
-            reply(client, interaction, 'No player lists cached. Please use \"/match <player count>\" instead');
+            reply(interaction, 'No player lists cached. Please use \"/match <player count>\" instead');
             return;
         }
 
@@ -39,7 +39,7 @@ module.exports = {
 
         // make the teams
         if (!helper.makeTeams(player_vals, message, message.client, data.stdev_ratio)) { // if teams aren't made, let them know
-            reply(client, interaction, 'Unable to make teams with these players. Sorry :(');
+            reply(interaction, 'Unable to make teams with these players. Sorry :(');
         }
     },
 };