+++ commands/users/v.js
@@ -1,7 +1,7 @@
 // js file for the viewing version command
 
 // info about the node.js project
-const { reply } = require('../../helper_functions/event_helper.js');
+const { reply } = require('../../helper_functions/comm_helper.js');
 const package = require('../../package.json');
 
 module.exports = {
@@ -17,6 +17,6 @@ module.exports = {
     // actual command code
 	async execute(interaction, args, client) {
         // get version from package file
-        reply(client, interaction, `MatchMaker v${package.version}`);
+        reply(interaction, `MatchMaker v${package.version}`);
     },
 };